# swagger.yaml
basePath: /
consumes:
    - application/json
definitions:
    ConfirmedBot:
        properties:
            detectionStatus:
                type: string
                enum:
                    - REPORTED
                    - IN_PROCESS
                    - BOT
                    - NOT_BOT
            botReason:
                type: string
                enum:
                    - BOT
                    - VIOLENCE
                    - FAKE
            platform:
                type: string
                enum:
                    - TWITTER
                    - FACEBOOK
                    - INSTAGRAM
        required:
            - detectionStatus
            - botReason
            - platform
        type: object
    Bots:
        properties: {}
        type: object
        additionalProperties:
            $ref: '#/definitions/ConfirmedBot'
    Report:
        properties:
            reporterKey:
                type: string
            platform:
                type: string
                enum:
                    - TWITTER
                    - FACEBOOK
                    - INSTAGRAM
            botReason:
                type: string
                enum:
                    - BOT
                    - VIOLENCE
                    - FAKE
            userId:
                type: string
            description:
                type: string
        required:
            - reporterKey
            - platform
            - botReason
            - userId
            - description
        type: object
info:
    title: 'Botim - extension backend'
    version: 1.0.0
    description: 'Using latest JavaScript features on the server'
    license:
        name: MIT
paths:
    /bots/confirmed:
        get:
            operationId: GetConfirmed
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/Bots'
                '501':
                    description: 'Server error'
            description: 'Get all confirmed bots.'
            tags:
                - Bots
            security: []
            parameters:
                -
                    in: query
                    name: userIds
                    required: true
                    collectionFormat: multi
                    type: array
                    items:
                        type: string
    /bots/suspected:
        post:
            operationId: ReportSuspected
            produces:
                - application/json
            responses:
                '204':
                    description: 'No content'
                '501':
                    description: 'Server error'
            description: 'Report a bot.'
            tags:
                - Bots
            security:
                -
                    reporterAuth: []
            parameters:
                -
                    in: body
                    name: report
                    required: true
                    schema:
                        $ref: '#/definitions/Report'
produces:
    - application/json
swagger: '2.0'
securityDefinitions:
    reporterAuth:
        type: apiKey
        name: Cookie
        in: header
host: 'localhost:8080'
tags:
    -
        name: Bots
        description: 'Botim on you, israel.'
schemes:
    - http